Add "dir" and "*.info" to the default :files spec
Since most packages do not have `dir` or `*.info` files, this required
changing the behaviour when a wildcard entry in `:file` expands to an
empty file list. Previously, an error would be raised in this case; now,
an error is raised only when the *entire* `:files` spec expands to an
empty file list.

The slight danger, then, is that certain build failures due to upstream
reorganisation of files might be less obvious.