Divide package-builder.el into more sections
The basic idea is, that by splitting the file into smaller parts,
which are ideally focused on one particular subject, it becomes easier
to find the parts that are relevant to the task at hand, and to get an
overall feel for the provided functionality and where the complexity
lies.  This commit does not achieve that, but it is a step in the
right direction.

Furthermore this makes it less likely that the already somewhat shaky
existing order of the library is further reduces by sloppy placement
of new definitions.  Placing a new definition "next to something that
is closely related (and calling it a day)" often is not good enough,
one also has to take the wider context and overall structure of the
file into account.  Such sloppiness accumulates and finer granularity
helps by discouraging it.

One instance of a badly placed addition is 8f76df9a, which placed a
section titled "Helpers for recipe authors" right into the middle of a
section titled "Public interface", ignoring the fact that not only the
code before where the new code was inserted belongs to that section,
but also the code that follows.  As a result many of the definitions
in "Helpers for recipe authors" have absolutely nothing to do with
that.

The name "Public interfaces" also was a bit unfortunate (and probably
encouraged misplacing "Helpers for recipe authors").  This commit
renames it to "Building" because that's what all the contained
definitions are about.  Some of them are part of the public interface
and some of them are very low-level.

Additionally a "second" section "Building" is added, after the "Helper
for recipe authors" section.  The next commit fixes this for good by
joining the two "Building" sections together, by moving the second one
before "Helper for recipe authors".