Remove unnecessary package-build--recipe-alist-initialized
We don't need an additional variable to track whether
`package-build--recipe-alist' has been initialized, because we can
just check whether the value of the latter is non-nil to do so.

Strictly speaking, this is not exactly equivalent, since it is also
possible that not a single recipe exists, while the alist has already
been initialized with the contents of "all of those" recipes.

We can simplify this anyway, because the cost of determining that
there are zero recipes is negligible and this package isn't very
useful in that case anyway.