package-build--checkout-git: cosmetics
It is not necessary to let-bind a variable here because its value
is used only once.  If we do use a variable, then it makes no sense
to only consider the first two out of three possible sources of the
value.  And even if we do that anyway then it still makes no sense
to split like so
  (let ((commit (or COMMIT BRANCH-A)))
    (or commit BRANCH-B))
instead of
  (let ((commit COMMIT))
    (or commit BRANCH-A BRANCH-B))
especially when we consider that for both BRANCH-A and BRANCH-B
we have to do (concat "origin/" BRANCH).