package-build--git-repo: use plumbing command to get value
Also use `process-lines' instead of `package-build--run-process-match'.
We can do that because the complete output is now the string that we
are interested.  Previously when we still used the wrong command, we
had to find the desired part among some noise.

Switching the function that runs the subprocess also has these effects:

1. We no longer make sure that the subprocess doesn't take too long
   to run.  But since that cannot happen for this command, that is
   not a problem.

2. We no longer show some custom error message when `default-directory'
   is not a valid directory.  But because the default error message
   that is shown instead is just as good, that is not a problem.