Consistently identify packages using their name, a string
For several functions change the name of the argument that represents
a package to "name".  Previously a mixture of "name", "package-name",
"pkg", and "file-name" was used.  In the last case the value had to be
a file.

For some of these functions also change the type of that argument to
be the name of a package, a string.  Likewise change the value
returned by some of these functions to be a package name.  These are
the functions whose argument or return value type changed:

* package-build-read--recipe: takes a package name as argument instead
  of a path to a recipe file.

  This affects package-build--read-recipes and
  package-build--read-recipes-ignore-errors.

* package-build--package-name-completing-read: returns a package name
  instead of a symbol.

  This affects package-build-archive and
  package-build-archive-ignore-errors.

* package-build-archive, package-build-archive-ignore-errors: take a
  package name as argument instead of a symbol.

  This affects package-build-current-recipe.

* package-build-package: takes a package name as argument.  Previously
  a symbol was also accepted.

  This affects package-build-archive.

* package-build-create-recipe: takes a package name as argument
  instead of a symbol.

  This affects no function.

* package-build--find-package-file: takes a package name as argument
  instead of a symbol.

So the only functions whose package argument and return value are not
changed, but whom are affected by such changes in other packages are:

* package-build--read-recipes
* package-build--read-recipes-ignore-errors
* package-build-current-recipe