package-build--run-process: add DESTINATION argument
Add a new argument DESTINATION, which controls where output goes.

If it is nil, then it goes into the buffer named
"*package-build-checkout*", which is very useful because previously
most callers had to make that buffer current to keep this function
from cluttering some buffer.

If it is t then output goes into the current buffer, which is useful
because for a few callers that is the right thing to do.

If it is an actual buffer, then output goes into that buffer, which
isn't really used, but makes this somewhat consistent with built-in
functions that run sub-processes.