package-build--{git,hg}-repo: rename to package-build--used-{git,hg}-url
REPO is ambiguous - it might refer to the local repository or to the
url of the remote repository "connected" to that repository.  URL on
the other hand is quite clear.

The "git"/"hg" part won't be needed anymore once we switch to using
eieio.  So you can ignore that noise.

The "used" is added to differentiate these functions from the method
`package-build--upstream-url', which will be added soon.

Also move the definitions of these functions.