Remove the WORKING-DIR or DIR argument of many functions
Instead of passing the value around, determine it once it is needed.
Add a new function `package-recipe--working-tree' to do so.  The name
of that function does not match the name of the library that defines
it because it will be moved to a library with a matching name in a few
commits.

Don't do this for a few functions that will be removed soon anyway.

In many cases the DIR argument is replaced with a new argument NAME
because `package-recipe--working-tree' needs to know for what package
it should return the working tree.  However in a later step that new
argument as well as the existing CONFIG argument will be replaced with
a single argument RCP.  What is being done here is but a step in that
direction.