package-build-archive: Dump archive when called interactively
Actually add a new argument DUMP-ARCHIVE-CONTENTS, which is
t when called interactively but could also be used otherwise.

When building a single package the archive-contents have to
be updated just like when building multiple packages.  In the
latter case we continue to do it only once when all packages
have been build for performance reasons.