make: Suppress obsolescence warnings
We are aware that these functions and macros are obsolete but we
cannot stop using them because we still support older Emacsen in
which their replacements do not yet exist.  By suppressing these
warnings we are once again able to spot other warnings that were
previously drowned out by all the warnings we had already seen.