package-build--archive-alist-for-json: Make data json-compatible
When using a newer Emacs, which defines `package-desc-create', then we
use some of information that is extracted by `package-buffer-info' but
that cannot turned into JSON as is.

This finally surfaced in https://github.com/melpa/melpa/issues/5584
after we finally updated Melpa to use such an Emacs.

The easiest and cleanest fix would be to just not use that information
and instead extract and format it ourselves.  However that would mean
that we would store information in archive-contents that does not
conform to what other packages who *do* use the functionality provided
by `package.el' itself would put there.

So we continue to use that data as-is and only modify it when
exporting to JSON.

The change to `pacakge-build--archive-entry' reverts the temporary
kludge introduced by befbd7fcb9d1c7143e3ce432eaae69160112e96d.