package-build-dump-archive-contents: Absorb package-build--archive-entries
Previously the existence of the latter function made it possible
to remove the obsolete entry files ("*.entry") and archive files
("*.tar" or "*.el") without also updating the "archive-contents"
file.  That is not actually desirable so don't support it anymore.

Define `package-build--archive-entries' as an alias for backward
compatibility, in case some third-party package calls it directly.