Bug was causing stable builder to delete newer versions.
Code in package-build-dump-archive-contents looks in entries for other
entry files with the same name. It calls this `other-entry`. At some
point the versions are compared and if `entry` has a lower version
than `other-entry`, the values of those variables are swapped. Later
on `other-entry` has its files deleted and it is removed from the
listof entries.

The code was a bit confused -- reasonably so as its a bit confusing --
and accidentally removed the files for `entry` instead of
`other-entry`.

It would happen that the first time this ran, the correct
archive-contents would be generated but the files would be deleted.
Subsequent runs would cause returning to the old version.

I suspect that a lot of stable packages are going to get updated when
this goes in. Although, it was dependent on how reading the entry
files happened. It's possible in some cases the larger version gets
read first and in that case this bug wouldn't show up.

fixes #6644