Remove some library header keywords
Remove the "Version" header keyword.  It was still stuck at 0.1 even
though we are at 2.5 -- go figure.

Remove the "Created" header keyword.  In the age of Git we can get
that sort of information using Git.  While the first commit was
created on 2011-10-05, work actually began five days earlier, but that
does that *really* matter eight years later?  I think not.