Use package-desc structs as-is and generate them ourselves
Previously we used `package-buffer-info' instead and then we converted
the returned struct into a less expressive vector.

We used to do the latter to maintain compatibility with "package.el"
from < Emacs 24.4, but since we now require >= Emacs 25.1 we can
finally stop doing that and embrace the new struct instead.

Also stop using `package-buffer-info' because that isn't and never was
suitable for our use-case.  It is less work to do everything ourselves
than to revert things we don't want.