package-build--desc-from-library: Take files alist as argument
Now the four first arguments are the same for this function and
for `package-build--desc-from-package', which will start using
this function in the next commit.  The latter means that one of
these to functions will have to start determining the library
based on NAME and FILES and it makes more sense if `*-library'
does it.

Also stop `ignore-errors' because it doesn't make sense to do so
here (when `package-build--desc-from-package' even raises errors
explicitly).  Sneak that into this commit because it makes for a
simpler diff.