package-build--checkout-1: Implement for package-hg-recipe
Unlike the respective method for `package-git-recipe' this method does
not re-checkout the tip of the main branch when no other revision is
given.  This is mainly because I don't know how to do that with hg.

We add this method so that the caller can be refactored without having
to worry about difference between the two vcs.