package-build-get-{tag,timestamp}-version: Consistent return value
All `package-build-get-*-version' must now return (REV . VERSION).
Currently REV isn't always used, but that is going to change.
Previously `package-build-get-tag-version' returned the tag instead
of a commit hash, but we don't actually care about the tag, just the
version string (which derives from the tag but may not be identical).