package-build-get-version-function: New option
Now that the user can specify an arbitrary function, we have
to be consistent about the order in which the current version
is determined and checked out.

Previously the order depended on `package-build-stable' and
while that made sense, it wasn't actually necessary because
`package-build-get-tag-version' does not actually need HEAD
to be checked out.

1. Call `package-build-get-version-function' to determine the
   version and the respective revision, as (REV . VERSION).
2. Check out that REV.  Often that is a no-op, but given an
   arbitrary function, we cannot know that ahead of time.
3. Return that VERSION.