package-build-get-timestamp-version(git): For correct commit
As a (somewhat questionable) optimization, for Git repositories we
initially only fetch without checking out the new head.  When building
a timestamp-based package we used to then immediately checkout the
latest commit on the appropriate branch.

In [1: 6fe53d8] we started to delay that checkout until after the call
to `package-build-get-timestamp-version' just as we already did when
building a stable package.

But we failed to adjust `package-build-get-timestamp-version' to take
into account that it can no longer rely on the last commit being
checked out.  We now do that by changing that function to operate on
"origin/HEAD" instead of whatever commit happens to be checked out.

We only do that for Git; for Mercurial this is unnecessary because in
that case we don't merely fetch, we pull.

1: 2021-07-05 6fe53d8e7691e2bbfd1de744154bf38dfb9c72c3
   package-build-get-version-function: New option