Perform checkout during version determination if necessary
All `package-build-get-version-function' are responsible for
determining the appropriate revision and the corresponding version
string.  In order to do the latter, they may have to checkout out
the revision.  This can be done using `package-build--checkout-1'.

Our `package-build-get-timestamp-version' for example doesn't call
that function directly, but it uses `package-build--get-timestamp',
which does.  (It would be better to make that unnecessary, but this
is how it currently works.)

Previously we considered the `commit' and `branch' slots only after
determining the version, but that is wrong because their values are
relevant when determining the version.  We move this logic into
`package-build-get-timestamp-version'.  Alternative implementations
will have to copy that logic, if they are supposed to respect these
slots as well (which isn't a given).  As before these slots are
ignored for mercurial (which ought to be fixed at some point).