Make sure the version regexp matches the tag
Before this change if the `string-match' failed, the call to
`match-string' would still return a result. A tag "0.16.0" with
a regexp of "foo/v\\(.*\\)" was failing to match but was still
return "0.16." as a result for the call to `match-string'.

This is because a failed `string-match' does not unset the match
data, instead it merely forgoes overriding the existing match
data:

  (progn
    (string-match "1" "1")
    (string-match "1" "2")
    (match-string 0 "3"))
  => "3"

With this change, `match-string' isn't even called if there is no
string match.