package-build--run-process: Fall back to correct default-directory
The intention is to fall back to the `default-directory' of the caller
when DIRECTORY is nil, but in the usual case when a persistent logging
buffer is used because DESTINATION is nil, then the local value of
that buffer is used but that is simply the value inherited from the
first caller.

Update the buffer-local value because that is nice when debugging, but
make sure only to do so when using the persistent buffer.  Continue to
let-bind the value for when the command is run in an arbitrary buffer.