package-build--generate-info-files: Simplify deleting texi file
In the past it was necessary to delay deleting the copied texi file
because it was used as the input of "texinfo".  We had to give up on
that because in some cases it is necessary to run that command in the
repository.

So now we can just delete the copied file immediately.