Add .dir-locals.el replacing duplicated settings
Like `lexical-binding, `coding' has to be set in each individual file,
so continue to do that.  I am unsure whether doing it at the end of
the file, as we did before, takes effect, so move that setting to the
first line.  Setting this should be unnecessary anyway, but since we
did it before, continue to do so anyway.