package-build-expand-files-spec: Replace similarly named function
The intention is to rename the existing function but because there are
downstream users of the old function we keep that around.  While this
commit only changes the function name, subsequent commits also change
the function signature, so we don't just use an obsolete function
alias but turn the old function into a wrapper around the new.

Also adjust the name argument and internal variable accordingly.

For the names of the new function and of its arguments and internal
variables, use the same logic as is already used for the variable
`package-build-default-files-spec': there is one spec, containing
multiple entries, specifying multiple files.

Also rename the DIR argument to REPO and the internal PREFIX variable
to SUBDIR.