Synchronize commit and timestamp information.
The functions package-build--get-commit and package-build--get-timestamp
are used to determine what snapshot is used for building a package
archive. However they use deviating methods to determine this
information resulting in conflicting information. Here we transplant the
mechanics from package-build--get-timestamp to package-build--get-commit
to establish a coherent state.

This is a step towards reproducible distfiles.