Use commit time of the commit actually used for mtimes in tarball
Recently we started using the existing function
`package-build--get-timestamp' to determine that time, but that is not
correct for releases.  We have to use the commit time of the commit
from which the tarball is build, but that function returns the time
of the most recent commit, reachable from that commit, that actually
touched any of the files that we are interested in.