Remove support for checking out submodules
This feature caused issue due to bitrot in packages that used it.
It also should not have been implemented in the first place as it
encourages bad practices.  Only two packages still depended on this
feature and they have been removed for the time being but should be
added back once the no longer depend a submodule.

I am the maintainer of one of these packages and expect to get my
act together soon. ;)