Remove faulty package-build-use-hg-purge option
When I added this option I made several mistakes, which canceled each
other out.  The docstring was misplaced, effectively always enabling
this feature.  The option was added under the assumption that this
extensions has to be explicitly enabled.  That is not the case; it is
a built-in extension and it is enabled by default.  The code used to
detect whether the extension is enabled was also faulty: for an
extension that is enabled by default, "hg config extensions.EXTENSION"
does not print an empty line, as it does for explicitly enabled third-
party extensions (so much for "hg's UI is more consistent than that of
git").

We could fix these issues and make this extension optional, just in
case someone actively *disables* it, but it is actually important
that this extension is available and nobody has reported that they
want to disable it, so we do the right thing instead and insist on
this being enabled.  If someone disables it, then they will get an
error and can talk to us.