Use git/hg to determine correct commit for snapshot
Before this commit we determined the files that match a package's
files-spec using an elisp implementation and then we used git/hg to
determine the latest commit in which one of these files was modified.
This had two issues:

1. It was necessary to check out the latest upstream commit as part
   of the commit selection step.

   This led to an unfortunate lack of separation of concerns, most
   of which were addressed in commits leading up to this one.

   The remaining lack of separation could be addressed by using a
   new elisp implementation, which operated on a git tree instead of
   on checked out files, but doing that would not address the second
   issue below.  Addressing that, would mean throwing out this new
   implementation in the very next commit, which is why this commit
   tackles these two issues at once.

2. The wrong commit was selected in at least the following edge-case:

   A first commit A touches a matching file.  A second commit B
   removes another matching file.  A third and final (HEAD) commit C
   doesn't modify any matching files, so no snapshot package is build
   for that commit.

   Never-the-less C is used to determine what the matching files are
   and because the file that was removed in B, no longer exists in C
   we concluded that A was the last commit in which something relevant
   happened.  The removal of a matching file in B warrants a new
   snapshot based on that commit, but because HEAD is at C, we used
   to built a snapshot for A instead.

The second issue could be addressed in elisp alone, i.e., we could
walk all commits without relying on git and hg, but that would be
very inefficient and likely error-prone.

This commit separates concerns between elisp and git/hg differently.
`package-build--spec-globs' converts the files spec into arguments
understood by git or hg.  Git or hg then takes care of finding the
last commit that touched any matching file.  This includes removals
of matching files.

Unfortunately both git and hg process all include rules before all
exclude rules, i.e., they disregard the order of the arguments.
Previously it was possible to override the default exclude rules in
the recipe-specific files spec, but we have to give up on that.  In
practice this change affects very few packages, three to be precise,
but it *is* a change in behavior, and it is forced upon us by this
limitation in git and hg.

Determining the latest commit that touches any of the matching files
and actually listing the matching files in that commit are separate
tasks, which were always performed separately.  It would have been
nice to also rely on git resp. hg to do the latter, but unfortunately
that is not possible.

While git and hg match the same files as the elisp implementation
(modulo the exclude complication), they do not produce the same
output.  When a directory matches a glob, then they list all files
within that directory instead of the directory itself.  That makes
sense from their perspective, but is not what we need.  So we still
need `package-build--expand-files-spec' (and have to adjust it to
process includes before excludes, to match the git/hg behavior).