package-build--create-tar: Add support for bsdtar
Gnu tar has an unfortunate feature that causes it to treat a path that
contains a colon as remote.  It has an option to disable that, but Bsd
tar does not have this feature and thus no option to disable it.  On
Windows this is a problem because absolute paths begin with the drive
letter followed by a colon.

Closes #34.