For hg, ensure non-nil timestamp and use specfied commit and branch
In order to get the latest revision set for a collection of files, the
revset to be searched must be sorted descending by date.  The revset
is constructed from all ancestors of a specific commit or the latest
commit of a branch, either specifically named or implicitely named
`default'.

See `hg help revsets'
(https://www.mercurial-scm.org/repo/hg/help/revsets).

Fixes #77.