Generate elpa-packages.eld for use by package-vc.el
`package-vc' expects GNU ELPA recipes as supported by `elpa-admin'.
For [Non]GNU ELPA, the "elpa-packages.eld" file distributed to
clients, is a reconstructed copy of the "elpa-package" file tracked
in the "main" branch of the [Non]GNU ELPA repository.

MELPA uses a different recipe format.  The main difference is that
MELPA includes only files specified in the recipe or the default file
specification in a package, FSF's ELPAs include everything that is not
explicitly excluded (and additionally encourages authors to exclude as
few files as possible).

`package' expects all elisp files and manuals to reside in the top-
level directory.  Both `elpa-admin' and `package-build' support moving
files there.  While `package' can rely on that being done by *ELPAs,
including MELPA; `package-vc' cannot do so.  Instead it uses the same
recipe as `elpa-admin'.

Since MELPA does not use the same recipe format, it cannot distribute
a fully [Non]GNU ELPA / `package-vc.el' compatible recipe.  We could
try to translate recipes, but that would not always work.  For now we
do not even attempt to do so.  That means that `package-vc' will most
likely fail for some packages, specifically those that put elisp files
in a subdirectory.