Include all maintainers in archive-contents
Set an additional property `:maintainers' to the list of maintainers,
but only if there actually are multiple maintainers.  Continue to set
`:maintainer' to the first of the maintainers.  We cannot use a list
of maintainers as the value of `:maintainer' because that would break
existing versions of Package.

We keep `:maintainer' for the benefit of older Package versions, and
we only set `:maintainers' if there are multiple maintainers to avoid
needlessly duplicating data.

Emacs 30.1 will be the first release to support displaying multiple
maintainers, see https://debbugs.gnu.org/cgi/bugreport.cgi?bug=62524.