Remove "-melpa" from some symbol names
For users of this library that aren't MELPA, that string is incorrect.
And for us it just needlessly makes the names longer, without any real
benefit.

Keep the name `package-build--melpa-base' because renaming that would
be more likely to cause breakage and, because it usually does refer
to the MELPA repository, even when used by something other than MELPA.