Remove package-build-write-badge-images variable
Instead allow `package-build-badge-data' to be nil, to disable the
creation of badges.  As before we default to not creating any badges.

`package-build-write-badge-images' should be set in the Makefile,
to the value appropriate for the channel that is being build.
Previously the text and color did not have to be specified in the
Makefile, but badge creation did have to be enabled; so this does
not really complicate the Makefile.

Also see previous commit.