No longer always signal an error if no version can be detected
On the "unstable" channel (and the upcoming "snapshot" channel) we
should always be able to come up with a version number, regardless of
whether upstream does releases.  If we fail to do so, then that might
be due to a bug in our code, or due to an upstream defect, and an
error is appropriate in either case.

On the "stable" channel we only publish packages whose upstreams tag
releases.  If they don't do that in a way we can detect, then that is
unfortunate but not necessarily a bug.  Don't signal an error in this
case, to avoid a very large number of false-positives.

On the upcoming "release" channel, we fall back to publishing the
same snapshot version as on the "snapshot" channel, when we cannot
determine the upstream release.  If we fail to either detect the
upstream release or else fall back to the snapshot version, then
that likely is a bug in our code, so an error is appropriate.