package-build--cleanup: Work around mercurial's broken purge command
I tried using these arguments:

  --all                  purge ignored files too
  -y --noninteractive    do not prompt, automatically pick the
                         first choice for all prompts

But sadly that does not work as advertised:

  $ hg purge --all -y
  permanently delete 1 ignored files? (yN) n
  abort: removal cancelled

  $ hg -y purge --all
  permanently delete 1 ignored files? (yN) n
  abort: removal cancelled

  $ hg purge --all --noninteractive
  permanently delete 1 ignored files? (yN) n
  abort: removal cancelled

Additionally there is no argument to include "unknown" files, i.e.,
files that are neither tracked nor ignored.  No such files should
exist in our case, because no human touches these repositories, but
they do.  So it appears that when one checks out a new revision that
lacks some files that the previously checked out revision had, these
files are left in the worktree.

So we have no choice but to manually clean out the ignored and unknown
files, and then run the purge command anyway, in case there happen to
be other files that have to be purged; and hope that it will comply
and remove those.