package-build-*-function: Remove experimental variables
This was useful during development and I considered sharing clones
between Melpa and the Emacsmirror locally.  The former has run its
course and the latter is not feasible because package-build.el
happily deletes clones, which would severely mess with emir.el.

These variables were always marked as experimental and users were
told to inform me if these were of use to them.