Store all extracted data in recipe object
Originally the recipe object only contained information from the recipe
file.  Then we started to collect additional information into the object
but that was not completed, we still collected some information into a
`package-desc' struct (the format used by package.el).  These structs
are inconvenient, because they have to remain compatible with older
package.el releases.

When *building* packages, there is no need to remain backward
compatible, so we can go away with this complication and store all
extracted information in the recipe object.  We still have to convert
to the nested structure expected by package.el, but now we do so at the
latest possible time, when writing the "*-pkg.el" files and the "entry"
files (which get collected into the "archive-contents" file.

- `package-build--write-archive-entry' and
  `package-build--write-pkg-file' now take a recipe object as argument
  instead of a package struct.  The new `package-build--extras' variable
  controls what "extra" information gets written to the generated files.

- Replace `package-build--desc-from-{library,package},
  with `package-build--extract-from-{library,package}.  While the old
  functions returned the extracted data and left it to the callers to
  store it in the struct, the new functions store the information in
  the recipe themselves.

- Reorder the slots in recipe objects, separating the slots that may be
  set from information in the recipe file, from the slots that are set
  to extracted information.  (The `commit' slot is a special-case, in
  very rare cases it is set in the recipe, but usually gets set later.)